First version of function for transforming active src-block headers.
Active means that src-block is in the same language as the associated
source-buffer. Transform means that header-args and switches are moved
from their position behind lang in the \#+begin_src line to separate
\#+header: lines on top of the src-block.

This way, header-args and switches become permanent even for those
src-blocks that will be transformed back to code when
the *outorg-edit-buffer* is closed.