Only run kill hooks for the outorg edit buffer.
When editing a file with org, temporary buffers are sometimes created
and killed for auxiliary processing. Unfortunately, when these temporary
buffers are killed, the outorg kill hooks were run, which reset various
pieces of global outorg state. This check only lets the kill hooks run
when the buffer being killed is the outorg edit buffer.