Support babel C++
This is a bit sketchy, but C++ and C are both supported by the same
babel library so we can not use the source block tag to determine which
babel library is required. If we think we need `cpp' we just look for
`C', since that is the library we actually need.